pyproject.toml
[tool.poetry]
name = "video-intelligence"
version = "0.1.0"
description = ""
authors = ["gagata <gagata@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
google-cloud-aiplatform = {extras = ["evaluation"], version = "^1.83.0"}
pillow = "^11.1.0"
tqdm = "^4.67.1"
fsspec = "^2025.2.0"
gcsfs = "^2025.2.0"
matplotlib = "^3.10.0"
opencv-contrib-python = "^4.11.0.86"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
google-genai = "^1.21.1"
label-studio = "^1.19.0"
db-dtypes = "^1.4.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
